op-acceptance: Move withdrawal test to own package.
Time travelling L1 a week into the future makes the L2 lag behind for a long time after the test completes. That causes unexpected flakiness.
Ultimately time travel tests need to have a disposable system.
