+++ op-acceptance-tests/tests/interop/proofs/withdrawal/init_test.go
@@ -1,4 +1,4 @@
-package proofs
+package withdrawal
 
 import (
 	"testing"
@@ -7,5 +7,5 @@ import (
 )
 
 func TestMain(m *testing.M) {
-	presets.DoMain(m, presets.WithSuperInterop())
+	presets.DoMain(m, presets.WithSuperInterop(), presets.WithTimeTravel())
 }
