+++ op-acceptance-tests/tests/interop/proofs/withdrawal/withdrawal_test.go
@@ -1,4 +1,4 @@
-package proofs
+package withdrawal
 
 import (
 	"testing"
